make follow reexports more efficient (vercel/turbo#7741)
### Description

When accessing many exports of a barrel file with many `export *` it was
a bit inefficient before since it was walking all star exports for every
accessed export.

After this change it only creates a map of the star exports once and
accessing every export only looks up the result in the map. This makes
it 4 times more efficient for a file with 140 `export *`. Probably even
more for bigger barrel files.

### Testing Instructions

<!--
  Give a quick description of steps to test your changes.
-->


Closes PACK-2747