chore: Remove unused `base16` dependency and dead `hex` wrapper function (vercel/turbo#8207)
### Description

Noticed these dependencies/functions were unused.

For the most part, it's better to do these conversions with the built-in
`format!()`.

Parsing base16 could make sense for this as a dependency, but we're not
doing that.

### Testing Instructions

```
cargo check
```

Also grepped through next.js to check that we're not using it there.